updated but wont post to database
