+++ app.js
@@ -64,7 +64,6 @@ app.use('/users', users);
 app.use('/chat', chat);
 
 
-
 // catch 404 and forward to error handler
 app.use(function(req, res, next) {
   var err = new Error('Not Found');
@@ -83,44 +82,5 @@ app.use(function(err, req, res, next) {
   res.status(err.status || 500);
   res.render('error');
 });
-
-// class methods
-
-/**
- * Ensure a user is logged in before allowing them to continue their request.
- *
- * If a user isn't logged in, they'll be redirected back to the login page.
- */
-requireLogin = function(req, res, next) {
-  if (!req.user) {
-    res.redirect('/login');
-  } else {
-    next();
-  }
-};
-
-/**
- * Given a user object:
- *
- *  - Store the user object as a req.user
- *  - Make the user object available to templates as #{user}
- *  - Set a session cookie with the user object
- *
- *  @ param {Object} req - The http request object.
- *  @ param {Object} res - The http response object.
- *  @ param {Object} user - A user object.
- */
-createUserSession = function(req, res, user) {
-  var cleanUser = {
-    firstName:  user.firstName,
-    lastName:   user.lastName,
-    email:      user.email,
-    data:       user.data || {},
-  };
-
-  req.session.user = cleanUser;
-  req.user = cleanUser;
-  res.locals.user = cleanUser;
-};
 
 module.exports = app;
