+++ controllers/users.js
@@ -1,7 +1,7 @@
 var express = require('express');
 var bcrypt = require('bcryptjs');
 var schema = require('../models/schema');
-var app = require('../app');
+var utils = require('./utils');
 var router = express.Router();
 
 // /* GET users listing. */
@@ -11,12 +11,7 @@ var router = express.Router();
 
 /* GET register page. */
 router.get('/register', function(req, res, next) {
-  res.render('register', { title: 'Chat It Up',
-                           pageName: 'Register',
-                           groupName: 'Project 2 : Group 3',
-                           user: 'user',
-                           csrfToken: req.csrfToken()
-                         });
+  res.render('register', { csrfToken: req.csrfToken() });
 });
 
 /* POST register from regisation form */
@@ -25,7 +20,7 @@ router.get('/register', function(req, res, next) {
  *
  * Once a user is logged in, they will be sent to the chat page.
  */
-router.post('/', function(req, res) {
+router.post('/register', function(req, res, next) {
   var salt = bcrypt.genSaltSync(10);
   var hash = bcrypt.hashSync(req.body.password, salt);
 
@@ -43,23 +38,17 @@ router.post('/', function(req, res) {
       if (err.code === 11000) {
         error = 'That email is already taken, please try another.';
       }
-
-      res.render('/user/register', { error: error });
-    } else {
-      app.createUserSession(req, res, user);
-      res.redirect('/chat');
+      return next(err);
     }
+    res.render('register', { error: error });
   });
+  utils.createUserSession(req, res, user);
+  res.redirect('/chat');
 });
 
 /* GET login page. */
 router.get('/login', function(req, res, next) {
-  res.render('/user/login', { title: 'Chat It Up',
-                        pageName: 'Login',
-                        groupName: 'Project 2 : Group 3',
-                        user: 'user',
-                        csrfToken: req.csrfToken()
-                      });
+  res.render('login', { csrfToken: req.csrfToken() });
 });
 
 /* POST login request */
@@ -68,16 +57,16 @@ router.get('/login', function(req, res, next) {
  *
  * Once a user is logged in, they will be sent to the chat page.
  */
-router.post('/login', function(req, res) {
+router.post('/login', function(req, res, next) {
   models.schema.findOne({ email: req.body.email }, 'fname lname username email password data', function(err, user) {
     if (!schema) {
-      res.render('/user/login', { error: "Incorrect email / password.", csrfToken: req.csrfToken() });
+      res.render('login', { error: "Incorrect email / password.", csrfToken: req.csrfToken() });
     } else {
       if (bcrypt.compareSync(req.body.password, user.password)) {
-        app.createUserSession(req, res, user);
+        utils.createUserSession(req, res, user);
         res.redirect('/chat');
       } else {
-        res.render('/chat', { error: "Incorrect email / password.", 
+        res.render('login', { error: "Incorrect email / password.", 
        });
       }
     }
