+++ middleware.js
@@ -1,5 +1,5 @@
 var models = require('./models/schema');
-var app = require('./app');
+var utils = require('./controllers/utils');
 
 /**
  * A simple authentication middleware for Express.
@@ -11,7 +11,7 @@ module.exports.simpleAuth = function(req, res, next) {
   if (req.session && req.session.user) {
     models.User.findOne({ email: req.session.user.email }, 'fname lname username email password data', function(err, user) {
       if (user) {
-        app.createUserSession(req, res, user);
+        utils.createUserSession(req, res, user);
       }
       next();
     });
