+++ views/window.pug
@@ -1,31 +1,29 @@
 extends layout
 
 block content
-    div.window
+    //- div.window
     
-    html(xmlns='http://www.w3.org/1999/xhtml')
-      head
-        title Chat - Customer Module
-        link(type='text/css', rel='stylesheet', href='/stylesheets/window.css')
-      #wrapper
-        #menu
-          p.welcome
-            | Chat It Up, 
-            b
-          p.logout
-            a#exit(href='#') Exit Chat
-          div(style='clear:both')
-        #chatbox
-        form(name='message', action='')
-          input#usermsg(name='usermsg', type='text', size='63')
-          input#submitmsg(name='submitmsg', type='submit', value='Send')
+    //- html(xmlns='http://www.w3.org/1999/xhtml')
+    //-   head
+    //-     title Chat - Customer Module
+    //-     link(type='text/css', rel='stylesheet', href='/stylesheets/window.css')
+    //-   #wrapper
+    //-     #menu
+    //-       p.welcome
+    //-         | Chat It Up, 
+    //-         b
+    //-       p.logout
+    //-         a#exit(href='#') Exit Chat
+    //-       div(style='clear:both')
+    //-     #chatbox
+    //-     form(name='message', action='')
+    //-       input#usermsg(name='usermsg', type='text', size='63')
+    //-       input#submitmsg(name='submitmsg', type='submit', value='Send')
 
-block scrpits
-  script(src='/socket.io/socket.io.js')
-  script.
-    var socket = io();
-    $('form').submit(function(){
-    socket.emit('chat message', $('#m').val());
-    $('#m').val('');
-    return false;
-    });
+    h2 User Information
+    p First Name: #{user.firstName}
+    p Last Name:  #{user.lastName}
+    p Email:      #{user.email}
+    p Data:       #{user.data}
+
+
